requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.100.2
SQLAlchemy==2.0.36
asyncpg==0.30.0
pandas==2.2.2
yfinance==0.2.52
langchain==0.3.7
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langchain-postgres==0.0.12
psycopg[binary]==3.2.3
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
